orders/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from .views import CartViewSet, OrderViewSet

router = DefaultRouter()
router.register("cart", CartViewSet, basename="cart")
router.register("orders", OrderViewSet, basename="orders")

urlpatterns = [path("", include(router.urls))]
